refactor: 로또번호 유효성 검사 메소드 책임 분리
