+++ src/utils/validator.js
@@ -8,6 +8,20 @@ const validator = {
       throw new Error(ERROR.NUMBER_INVALID_UNIT);
     }
   },
+
+  isDupplicateLottoNumber(numbers) {
+    const tempArray = [];
+    for (let i = 0; i < numbers.length; i++) {
+      if (tempArray.includes(numbers[i])) throw new Error(ERROR.NUMBER_DUPPLICATE);
+      else tempArray.push(numbers[i]);
+    }
+  },
+
+  isValidRangeNumber(numbers) {
+    numbers.forEach((number) => {
+      if (number < 1 || number > 45) throw new Error(ERROR.NUMBER_RANGE);
+    });
+  },
 };
 
 export default validator;
